Se agrega validación al delete de que se borre algo solo si existe

+++ src/controllers/student.controller.ts
@@ -301,10 +301,12 @@ export class StudentController {
         try {
             const dni = [req.params.dni];
             const response = await Queries.simpleSelectWithParameterContinous(query, dni, fetch, client);
-            const p = response.rows[0].file_path;
-            let fullPath = path.join(__dirname + '../../..' + '/public/' + p);
-            fs.unlinkSync(fullPath);
-            await Queries.simpleTransaction(deleteD, dni, client);
+            if(!response.rows[0] === undefined){
+                const p = response.rows[0].file_path;
+                let fullPath = path.join(__dirname + '../../..' + '/public/' + p);
+                fs.unlinkSync(fullPath);
+                await Queries.simpleTransaction(deleteD, dni, client);
+            }
             return res.status(200).json(
                 {
                     msg: 'CV deleted'
